Fix licensing issues
Switch license to GPLv2. As some of the authors did not reply to an
email asking whether they agreed or not on the change of license on
their work, just remove the files they created.

Most of subsequent work on these files has been performed by scripting
and can be replayed on demand.

Signed-off-by: Jean-Noël Avila <jn.avila@free.fr>