Remove es_MX translation
It is a huge burden for volunteers to try to fully translate a project
this size two times. There is nothing in Git’s terminology that merits
the creation of a duplicate Spanish translation. In Spanish-language
localization of computer software, dialectal files are mostly to address
minor vocabulary differences in colloquial language, meaning that locales
like es_MX or es_ES should only be produced for video games and related
programs, but not projects like Git.