Do not mis-spell the `git fsmonitor--daemon` command
Even in translations, the command names need to refer to the actual
Git command whose name is not to be translated.

The problem here was not so much a typo but an incomplete substition of
`--` by `{litdd}` in b5bb0fe (Translated using Weblate (Portuguese
(Brazil)), 2022-06-26).

This was reported by Todd Zullinger as a broken link to the manual page
in https://lore.kernel.org/git/ZWVIRXodL9pQZbtj@pobox.com/.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>