zh_HANS-CN: do separate hyperlink from the next word with a space
Git's website at https://git-scm.com/ uses Hugo to prepare the HTML
pages, and Hugo tries to translate bare HTTP/HTTPS links to proper HTML
links. It apparently uses similar heuristics to human beings when
determining where a link ends and the next word starts.

For example, it considers "https://git-scm.com/docs上查看。" to be a
hyperlink followed by a full stop. That is, it considers "上查看" _part_
of the hyperlink because there is no space or any other separator
between the _actual_ hyperlink and the following text. This part of the
translation was introduced in 8b72a13 (Translated using Weblate (Chinese
(Simplified)), 2023-02-11) and has been subsequently touched by 8b73fac
(mass type fixups in constant strings, 2023-12-21), 51ad8a5 (run
check_po script to find discrepancies in the po files, 2024-02-07), and
fcb0f4e (Translated using Weblate (Chinese (Simplified)), 2024-02-08),
neither of which inserted a space between the hyperlink and the
following text.

As a consequence, the last hyperlink in the section at
https://git-scm.com/docs/git/zh_HANS-CN#_%E6%8F%8F%E8%BF%B0 is
incorrect.

Let's rectify this.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>