git v2.49: update po files with fixed synopsis entries
The po4a tool used to extract the strings from the
documentation files was not able to handle the synopsis
entries. This caused the translation entries to be
on a single line instead of being formatted as verbatim.

Note that this version cannot be processed correctly with
the latest official po4a version. The po4a tool has been
patched locally, and the patch will be submitted to the
upstream repository.

Signed-off-by: Jean-Noël Avila <jn.avila@free.fr>